pyproject.toml
[tool.poetry]
name = "django-typer"
version = "2.1.0"
description = "Use Typer to define the CLI for your Django management commands."
authors = ["Brian Kohan <bckohan@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bckohan/django-typer"
homepage = "https://django-typer.readthedocs.io"
keywords = ["django", "CLI", "management", "Typer", "commands"]
classifiers = [
    "Environment :: Console",
    "Framework :: Django",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

packages = [
    { include = "django_typer" }
]
exclude = ["django_typer/locale"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
Django = ">=3.2,<6.0"
click = "^8.1.0"

# typer's release history is full of breaking changes for minor versions
# given the reliance on some of its private internals we peg the typer
# version very strictly to bug fix releases for specific minor lines.
typer-slim = ">=0.12.0,<0.13.0"

# this should track typer's rich dependency, so long as our console
# patches still work - so be sure to test on the low end of the range
rich = { version = ">=10.11.0,<14.0.0", optional = true }

shellingham = ">=1.5.4"

# we need this on 3.8 for Annotated types and 3.9 for ParamSpec
typing-extensions = { version = ">=3.7.4.3", markers = "python_version < '3.10'" }


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
Sphinx = [
    { version = "^7.2.0", markers = "python_version > '3.8'" },
    { version = "^7.0.0", markers = "python_version <= '3.8'" },
]
sphinx-rtd-theme = "^2.0.0"
mypy = "^1.0"
pylint = "^3.0"
doc8 = "^1.1.1"
aiohttp = "^3.9.1"
readme-renderer = {extras = ["md"], version = ">=42,<44"}
sphinxcontrib-typer = {extras = ["html", "pdf", "png"], version = "^0.3.0", markers="python_version >= '3.9'"}
scikit-learn = "^1.0.0"
pytest-env = "^1.0.0"
numpy = [
    { version = ">=1.26", markers = "python_version > '3.8'" },
    { version = "<=1.24", markers = "python_version <= '3.8'" },
]
scipy = [
    { version = ">=1.11", markers = "python_version > '3.8'" },
    { version = "<=1.10", markers = "python_version <= '3.8'" },
]
django-stubs = "^4.2.7"
pexpect = "^4.9.0"
pyright = "^1.1.357"
ruff = "^0.4.1"
graphviz = "^0.20.3"
sphinx-tabs = "^3.4.5"

[tool.poetry.extras]
rich = ["rich"]

[tool.mypy]
# The mypy configurations: http://bit.ly/2zEl9WI
allow_redefinition = false
check_untyped_defs = true
disallow_untyped_decorators = false
disallow_any_explicit = false
disallow_any_generics = false
disallow_untyped_calls = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
local_partial_types = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
exclude = ["tests"]


# todo doc8 configuration here is not being picked up and doesnt seem to be working
# as expected - for now pass these parameters manually
[tool.doc8]
max-line-length = 100
sphinx = true

[tool.pytest.ini_options]
# py.test options:
DJANGO_SETTINGS_MODULE = "tests.settings.base"
python_files = "test_*.py"
norecursedirs = "*.egg .eggs dist build docs .tox .git __pycache__ shellcompletion"
env = [
    "TERMINAL_WIDTH=80",
]

addopts = [
  "--strict-markers",
  "--cov=django_typer",
  "--cov-branch",
  "--cov-report=term-missing:skip-covered"
]

[tool.coverage.run]
branch = true
source = ["django_typer"]
concurrency = ["multiprocessing"]
parallel = true
relative_files = true
command_line = "-m pytest --cov=django_typer"

[tool.coverage.paths]
source = [
    "django_typer"
]

[pylint]
output-format = "colorized"
max-line-length = 88

[tool.pylint.'DESIGN']
max-args=30
max-locals=30  # there are lots of options!
max-module-lines=2000
max-attributes=20
max-statements=100

[tool.pylint.'MASTER']
ignore="tests"

[tool.pylint.'MESSAGES CONTROL']
disable = [
    'R0903', # too few public methods - seriously?
    'W0613', # unused argument - unavoidable in generic base interfaces
    'W0212', # access to a protected member of a client class - lots of monkey patches required here
    'C0415', # imports outside top level. I put them where they need to be!
    'C0413', # wrong import order. I put them where they need to be!
    'C0411', # wrong import order. I put them where they need to be!
    'W0603', # Using the global statement
]

[tool.pyright]
exclude = ["tests/**/*"]
include = [
    "django_typer"
]

[tool.ruff]
line-length = 88
exclude = [
    "doc",
    "dist",
    "examples"
]

[tool.ruff.lint]
exclude = [
    "tests/**/*"
]
